refactor: 훅 통합
