+++ src/app/product/[id]/hooks/index.ts
@@ -1,2 +1,1 @@
 export { default as useSelectedOptions } from './useSelectedOptions';
-export { default as useCartManager } from './useCartManager';
